Fix error messages for wallet and inventory.

+++ esop_trading/src/main/kotlin/Models/AddToInventoryInput.kt
@@ -1,3 +1,5 @@
 package models
 
-data class AddToInventoryInput(val esop_type: String?, val quantity:Int? =null)
+import java.math.BigInteger
+
+data class AddToInventoryInput(val esop_type: String?, val quantity: BigInteger? =null)
